Changed CPU usage percentage value type from integer into floating point (double precision).

+++ src/utils/cpuusage.h
@@ -4,7 +4,7 @@
 #include "procstat.h"
 
 
-typedef unsigned PercentageValue_t;
+typedef double PercentageValue_t;
 
 
 typedef struct CpuUsageInfo
